refactor: ♻️ rename errorUtil to httpErrorUtil and rehome parseErrorV2
What survives in this file — getAxiosErrorMessage and
getArchivedTeamConflictMessage — is purely HTTP-domain, and that is
exactly what makes it not redundant with classifyError: an AxiosError is
not a viem BaseError, so classifyError would bucket it as 'unknown' and
surface "Request failed with status code 409". The name now says which
of the two you want.

parseErrorV2 moves to generalUtil, next to log. It was never HTTP —
it condenses any Error to one line, and both its call sites use it
solely as a log argument. Leaving it behind would have made the new
filename a lie.

The contracts README documented parseError as the tool for non-V3 error
paths; that guidance now points at classifyError, and the log bullet
says to log the raw error rather than a pre-stringified one.

Closes #2395
