+++ app/src/components/sections/DashboardView/DeleteMemberModal.vue
@@ -66,7 +66,7 @@
 <script setup lang="ts">
 import { computed, ref } from 'vue'
 import { useTeamWriteGuard } from '@/composables/useTeamWriteGuard'
-import { getAxiosErrorMessage } from '@/utils/errorUtil'
+import { getAxiosErrorMessage } from '@/utils/httpErrorUtil'
 import type { Member } from '@/types'
 import { useDeleteMemberMutation } from '@/queries/member.queries'
 import UserComponent from '@/components/UserComponent.vue'
