+++ app/src/utils/generalUtil.ts
@@ -33,6 +33,19 @@ export const log = {
 }
 
 /**
+ * Condenses an Error into a single log-friendly line: its name plus the first
+ * sentence of its message. Lives beside `log` because that is its only use —
+ * for anything a user reads, reach for `classifyError` instead.
+ *
+ * @returns Error Name + First sentence of Error Message
+ */
+export const parseErrorV2 = (error: Error) => {
+  const message = error.message || 'Unknown error'
+  const firstSentence = message.includes('.') ? message.split('.')[0] : message
+  return `${error.name}: ${firstSentence}`
+}
+
+/**
  * @deprecated Import `formatAddress` from `@/utils/format` instead — this is
  * the same helper under an older name, kept while call sites migrate.
  */
