+++ app/src/utils/index.ts
@@ -1,7 +1,7 @@
 export * from './communityCreditOfferUtil'
 export * from './communityCreditUtil'
 export * from './communityCreditWizardUtil'
-export * from './errorUtil'
+export * from './httpErrorUtil'
 export * from './classifyError'
 export * from './generalUtil'
 export * from './constantUtil'
